+++ src/createStore.ts
@@ -1,10 +1,7 @@
 import { Store, Reducer, getId, Action } from './model'
 import { getState as _getState, map } from './createReducer'
 
-const subscribersEmpty = {
-  forEach() {},
-}
-
+// FIXME: add middleware
 export function createStore<R extends Reducer<any>>(
   rootReducer: R,
   preloadedState = null,
@@ -25,18 +22,39 @@ export function createStore<R extends Reducer<any>>(
     return state
   }
 
+  // FIXME: add ensureCanMutateNextListeners
   function subscribe<T>(
     listener: (a: T) => any,
     target: Reducer<T> = localReducer,
   ) {
     const targetId = getId(target)
-    if (subscribers[targetId] === undefined) {
-      subscribers[targetId] = new Set()
+    let subscribersById = subscribers[targetId]
+
+    if (subscribersById === undefined) {
+      subscribersById = subscribers[targetId] = new Set()
+      if (state.flat[targetId] === undefined) {
+        const {
+          _types,
+          _node: { _children },
+        } = localReducer
+        _children.push(target._node)
+      }
     }
 
-    subscribers[targetId].add(listener)
+    subscribersById.add(listener)
 
-    return () => subscribers[targetId].delete(listener)
+    if (initialState.flat[targetId] === undefined) {
+      return () => {
+        subscribersById.delete(listener)
+        if (subscribersById.size === 0) {
+          const { _children } = localReducer._node
+          delete subscribers[targetId]
+          _children.splice(_children.indexOf(target._node), 1)
+        }
+      }
+    }
+
+    return () => subscribersById.delete(listener)
   }
 
   function dispatch(action: Action<any>) {
@@ -44,15 +62,17 @@ export function createStore<R extends Reducer<any>>(
     if (newState !== state) {
       state = newState
 
-      state.changes.forEach(id =>
-        (subscribers[id] || subscribersEmpty).forEach(listener =>
-          listener(state.flat[id]),
-        ),
-      )
+      state.changes.forEach(id => {
+        const subscribersById = subscribers[id]
+        if (subscribersById !== undefined) {
+          subscribersById.forEach(listener => listener(state.flat[id]))
+        }
+      })
     }
     return state.root
   }
 
+  // FIXME: add replaceReducer
   return {
     subscribe,
     getState,