+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-<!-- ## 0.0.2 -->
+## 1.6.0
+
+Add lazy reducers to createStore
 
 ## 1.5.0
 
-Rewrite core - improve bundle size
+Rewrite core - improve bundle size and performance
 
 ## 1.4.0